+++ config/SZBE69/splits.txt
@@ -16,8 +16,8 @@ Sections:
 rb3/main.cpp:
 	.text       start:0x8000F96C end:0x8000F9B0
 
-rb3/string.cpp:
-	.text       start:0x80361F88 end:0x803620B4
+unknown/string.cpp:
+	.text       start:0x80361B10 end:0x803620B4
 
 MSL_C/mem.c:
 	.text       start:0x806E7E94 end:0x806E8004
