--- src/rb3/String.hpp
@@ -1,16 +1,0 @@
-#include <types.h>
-
-class String {
-public:
-    int len;
-    char * text;
-
-    String();
-    String(char * src);
-    String(String * src);
-    ~String();
-
-    void fn_80361F88(String * src); // clone?
-    void operator_=(char * src);
-    bool operator_==(char * src); // TODO match this
-};
