+++ src/unknown/unknown_prototypes.hpp
@@ -1,0 +1,4 @@
+char * fn_8000DB9C(String * str);
+void fn_80361F04(String * str, char * src);
+void fn_80362260(String * str, int len); // no idea
+void fn_80361C90(String * str, int len);
